+++ wasi/src/test/resources/compiled/greet-wasi.rs.wasm
