+++ wasi/src/test/resources/compiled/hello-wasi.rs.wasm
