+++ wasi/src/test/resources/compiled/hello-wasi.wat.wasm
