+++ wasi/src/test/resources/sources/greet-wasi.rs
