+++ wasi/src/test/resources/sources/hello-wasi.rs
