+++ wasi/src/test/resources/sources/hello-wasi.wat
